SLING-4957 - Sightly RenderContextImpl contains utility methods that don't belong there - Moved utility methods from RenderContextImpl to RenderUtils - Made methods static - Removed unnecessary passing of renderContext just for utility methods
