+++ bundles/scripting/sightly/engine/src/test/java/org/apache/sling/scripting/sightly/impl/engine/runtime/RenderContextImplTest.java
@@ -18,7 +18,6 @@
  ******************************************************************************/
 package org.apache.sling.scripting.sightly.impl.engine.runtime;
 
-import java.util.Collection;
 import java.util.HashMap;
 import java.util.Map;
 
@@ -27,8 +26,6 @@ import javax.script.SimpleBindings;
 
 import org.apache.sling.api.resource.ResourceResolver;
 import org.apache.sling.scripting.sightly.extension.RuntimeExtension;
-import org.apache.sling.scripting.sightly.impl.engine.runtime.RenderContextImpl;
-import org.apache.sling.scripting.sightly.impl.utils.RenderUtils;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
@@ -64,13 +61,4 @@ public class RenderContextImplTest {
         assertNotNull("Expected a non-null resource resolver.", scriptResolver);
     }
 
-    @Test
-    public void testGetCollectionWithOneElement() {
-        String stringObject = "test";
-        Integer numberObject = 1;
-        Collection stringCollection = RenderUtils.toCollection(stringObject);
-        assertTrue(stringCollection.size() == 1 && stringCollection.contains(stringObject));
-        Collection numberCollection = RenderUtils.toCollection(numberObject);
-        assertTrue(numberCollection.size() == 1 && numberCollection.contains(numberObject));
-    }
 }
