+++ bundles/scripting/sightly/testing/src/main/provisioning/model.txt
@@ -38,9 +38,6 @@
     org.apache.sling.commons.log.file="logs/error.log"
     org.apache.sling.commons.log.level="error"
 
-  org.apache.sling.scripting.sightly.impl.engine.SightlyEngineConfiguration
-    org.apache.sling.scripting.sightly.devmode=B"true"
-
 [settings]
   sling.system.packages.class.com.vladium.emma.rt.RT=com.vladium.emma.rt
   org.apache.sling.commons.log.julenabled=true
